pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pandas_wrapper"
version = "0.0.2"
authors = [
    {name = "Luis Teodoro"},
]
description = "A simple python package that integrate pandas with other tools."
requires-python = ">=3.10"
dependencies = [
    "pandas==2.2.2",
    "psycopg2-binary==2.9.9",
    "sqlalchemy==2.0.32",
]

[tool.hatch.build.targets.wheel]
packages = ["pandas_wrapper"]
